[nomaster] SI-5353 prohibits Array()
TODO: I'm about to fix SI-5923, and that fix is going to uncover SI-5353
(see more information here: https://github.com/scala/scala/pull/2464).

Since the fix to SI-5353 is going to bring binary incompatibilities,
we cannot include it into 2.10.x, but we want to fix SI-5923 in 2.10.2.

Therefore we decided to emulate the portion of SI-5923 that used to mask
SI-5353. That's only a temporary hack targeted at 2.10.x. There's no
reason for this code to be merged into master.