toolboxes now use empty package
To eval code, toolboxes need to wrap it into something tangible, because
that code could be just an expression that doesn't make sense to the
compiler on its own.

The wrapper includes a module and a method inside that module, s.t.
we can just compile it and then reflectively get an instance of the module
and just call its method.

Previously we used to wrap the module in a package def with a synthetic
name. But, on the one hand, this doesn't buy us anything, and, on the
other hand, this prevents the evaluated code from seeing identifiers
declared in the empty package. Looks like removing the synthetic package
from the equation is a clear win.